maybe fix bug

+++ src/Arp.h
@@ -134,7 +134,7 @@ class Arp {
   byte sequenceStepVelocity[MAX_STEPS_IN_SEQ] = {0, 32, 64, 32};
 
   // Total sequence length in discrete steps
-  byte totalSequenceSteps = 4;
+  uint16_t totalSequenceSteps = 4;
   // Total sequence length in pulses
   uint16_t totalSequenceLength = 4 * PULSES_PER_QUARTER_NOTE;
